daemon: Set content protection through IAS compositor
This change allows the daemon to talk with IAS Wayland compositor
to set the content protection property. It does so by dynamically
loading a library called wl_base and then making the
set_content_protection call into this library.
Note that this library must be present in the same directory as the
daemon. wl_base library is built with the IAS compositor and can be
found here: https://github.com/intel/ias

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>